+++ pkgs/objective_c/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add various ObjC categories (extension methods) to the built in classes.
 - Add various ObjC protocols to the bindings.
 - Make all visible API types public.
+- Add a `osVersion` getter, which returns the current MacOS/iOS version.
 
 ## 4.1.0